+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Extract `AbstractStandardProcessor.errorIfProcessingOverAndInvalidTypesDetected()` to simplify reuse in subclasses.
 * Expose `ElementsUtil.getTopLevelElement(Element)` that returns the top-level class, interface, enum, etc within a package that contained the specified `Element`.
 * Expose `AbstractStandardProcessor.processTypeElements(...)` helper methods to make subclassing easier.
+* Remove `AbstractStandardProcessor.process()` method and instead require that it is implemented in subclasses.
 
 ### [v0.08](https://github.com/realityforge/proton/tree/v0.08) (2019-12-30) · [Full Changelog](https://github.com/realityforge/proton/compare/v0.07...v0.08)